API Change: reorder `inflateAndSetViewModel`, move it and setContent view to databinding package
