+++ SecretSauce/src/main/java/com/byoutline/secretsauce/databinding/DataBindingHelper.kt
@@ -1,9 +1,12 @@
 package com.byoutline.secretsauce.databinding
 
-import android.databinding.Observable
-import android.databinding.ObservableBoolean
-import android.databinding.ObservableField
-import android.databinding.PropertyChangeRegistry
+import android.arch.lifecycle.ViewModel
+import android.databinding.*
+import android.support.annotation.LayoutRes
+import android.support.v4.app.FragmentActivity
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import com.byoutline.secretsauce.SecretSauceSettings
 import kotlin.properties.Delegates
 import kotlin.properties.ReadWriteProperty
 
@@ -89,6 +92,7 @@ object DataBindingHelper {
                 if (sender !is ObservableField<*>) {
                     return
                 }
+                @Suppress("UNCHECKED_CAST")
                 val obs: ObservableField<T> = sender as ObservableField<T>
                 val state = obs.get()
                 callback(state)
@@ -96,3 +100,41 @@ object DataBindingHelper {
         }
     }
 }
+
+
+/**
+ * Inflates view via [BINDING] class and sets [viewModel] into given [brVariableId].
+ * This method uses [SecretSauceSettings.brViewModelId] as default argument for variable.
+ * It is advice to set it globally in [SecretSauceSettings] (although you can provide this value
+ * as argument, it makes method call long enough that it may not be worth using)
+ *
+ * Example:
+ * ```
+ * private lateinit var binding: FragmentProjectsBinding
+ *
+ * binding = inflateAndSetViewModel(inflater, container, R.layout.fragment_projects, viewModel)
+ * ```
+ * or
+ * ```
+ * val binding: FragmentProjectsBinding = inflateAndSetViewModel(inflater, container, R.layout.fragment_projects, viewModel)
+ * ```
+ * Note: defining bidding type next to definition is encouraged. Otherwise you may need to specify this
+ * type in generic parameters instead.
+ */
+fun <VIEWMODEL : ViewModel, BINDING : ViewDataBinding> inflateAndSetViewModel(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        @LayoutRes layoutId: Int,
+        viewModel: VIEWMODEL,
+        brVariableId: Int = SecretSauceSettings.brViewModelId,
+        attachToParent: Boolean = false): BINDING {
+    val binding = DataBindingUtil.inflate<BINDING>(inflater, layoutId, container, attachToParent)
+    binding.setVariable(brVariableId, viewModel)
+    return binding
+}
+
+/**
+ * Alias for [DataBindingUtil.setContentView]
+ */
+fun <BINDING : ViewDataBinding> FragmentActivity.bindContentView(@LayoutRes layoutId: Int): BINDING =
+        DataBindingUtil.setContentView(this, layoutId)
