+++ sampleapplication/src/main/java/com/byoutline/sampleapplication/networkchangereceiver/NetworkChangeActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.byoutline.sampleapplication.R
 import com.byoutline.sampleapplication.databinding.NetworkActivityBinding
-import com.byoutline.secretsauce.lifecycle.bindContentView
+import com.byoutline.secretsauce.databinding.bindContentView
 import com.byoutline.secretsauce.lifecycle.lazyViewModelWithAutoLifecycle
 import com.byoutline.secretsauce.utils.NetworkChangeViewModel
