+++ sampleapplication/src/main/java/com/byoutline/sampleapplication/rx/RxLifecycleActivity.kt
@@ -5,8 +5,8 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.byoutline.sampleapplication.R
 import com.byoutline.sampleapplication.databinding.RxlifecycleActivityBinding
+import com.byoutline.secretsauce.databinding.bindContentView
 import com.byoutline.secretsauce.lifecycle.AttachableViewModelRx
-import com.byoutline.secretsauce.lifecycle.bindContentView
 import com.byoutline.secretsauce.lifecycle.lazyViewModelWithAutoLifecycle
 import java.util.concurrent.TimeUnit
 import io.reactivex.Observable as RxObservable
